nu_scaler_py/pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "nu_scaler"
version = "0.1.0"
description = "Hybrid Rust+Python real-time upscaling application."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PySide6",
    "Pillow"
]
# dependencies = []  # Add Python dependencies here
